move crab-agnostic code to its place
